Split RDW header primitives into dedicated microcrate
