+++ copybook-rdw-header/Cargo.toml
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 [package]
-name = "copybook-rdw"
+name = "copybook-rdw-header"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -8,8 +8,8 @@ license.workspace = true
 authors.workspace = true
 repository.workspace = true
 homepage.workspace = true
-description = "RDW framing microcrate with header primitives, reader/writer, and buffered helpers"
-documentation = "https://docs.rs/copybook-rdw"
+description = "RDW header primitives and buffer-level helpers"
+documentation = "https://docs.rs/copybook-rdw-header"
 readme = "README.md"
 keywords = ["cobol", "rdw", "record", "framing"]
 categories = ["encoding", "parsing", "data-structures"]
@@ -21,10 +21,7 @@ all-features = false
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
-copybook-core.workspace = true
 copybook-error.workspace = true
-copybook-rdw-header.workspace = true
-tracing.workspace = true
 
 [dev-dependencies]
 proptest.workspace = true
